backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
slowapi==0.1.9
redis==5.0.1
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
structlog==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6